beginning for bluetooth
comitting progress think im neirly finished, reconnecting after sleep is not perfect yet
more progress

more progress
idk some tryouts with settings

some env variables trying out

+++ skeleton/SYSTEM/tg5040/paks/MinUI.pak/launch.sh
@@ -104,17 +104,33 @@ CPU_PATH=/sys/devices/system/cpu/cpu0/cpufreq/scaling_setspeed
 CPU_SPEED_PERF=2000000
 echo $CPU_SPEED_PERF > $CPU_PATH
 
-# bt handling (todo, off for now)
-rfkill block bluetooth
-killall MtpDaemon # I dont think we need to micro manage this one
-
 # BT handling
 # on by default, disable based on systemval setting
-#bton=`/usr/trimui/bin/systemval bluetooth`
-#if [ "$bton" != "1" ] ; then
-#	/etc/bluetooth/bluetoothd start
-#	/usr/bin/bluealsa -p a2dp-source&
-#	touch /tmp/bluetooth_ready
+bton=`/usr/trimui/bin/systemval bluetooth`
+if [ "$bton" != "1" ] ; then
+	rfkill block bluetooth
+	killall MtpDaemon # I dont think we need to micro manage this one
+else
+	/etc/bluetooth/bluetoothd start
+	/usr/bin/bluealsa -p a2dp-source&
+	touch /tmp/bluetooth_ready
+
+
+	# Start Bluetooth daemon 
+	/etc/bluetooth/bluetoothd start
+	sleep 3
+
+	export BLUEALSA_TRANSPORT="a2dp-source"
+	export BLUEALSA_PCM_BUFFER_SIZE=4048
+	export BLUEALSA_VOLUME_CTL="softvol"
+	export BLUEALSA_LATENCY=100
+	export BLUEALSA_QUALITY="low"
+	# Start bluealsa in the background and log output
+	bluealsa > /mnt/SDCARD/bluealsa.txt 2>&1 &
+	# Turn Bluetooth power on
+	bluetoothctl power on
+	# Enable scanning for bluetooth devices
+	bluetoothctl scan on > /mnt/SDCARD/bluetoothctl.txt 2>&1 &
 #fi
 
 # wifi handling
